refactor: adjust visibility of UI module components
