+++ ui/src/main/java/com/example/ui/screens/search/SearchScreenRoute.kt
@@ -4,7 +4,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.example.ui.navigation.Route
 
-fun NavGraphBuilder.searchScreenRoute(){
+internal fun NavGraphBuilder.searchScreenRoute(){
     composable<Route.Search> {
         SearchScreen()
     }
