+++ ui/src/main/java/com/example/ui/screens/search/sections/filterDialog/genre/GenreUiModel.kt
@@ -3,7 +3,7 @@ package com.example.ui.screens.search.sections.filterDialog.genre
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 
-data class GenreUiModel(
+internal data class GenreUiModel(
     @DrawableRes val icon: Int,
     @StringRes val displayableName: Int
 )
