+++ viewModel/src/main/java/com/example/viewmodel/search/searchByKeyword/SearchUiState.kt
@@ -3,7 +3,7 @@ package com.example.viewmodel.search.searchByKeyword
 import com.example.viewmodel.common.MediaItemUiState
 import com.example.viewmodel.search.searchByKeyword.genre.MovieGenreItemUiState
 import com.example.viewmodel.search.searchByKeyword.genre.Selectable
-import com.example.viewmodel.common.TabOption
+import com.example.viewmodel.search.searchByKeyword.TabOption
 import com.example.viewmodel.search.searchByKeyword.genre.TvGenreItemUiState
 import com.example.viewmodel.search.searchByKeyword.genre.MovieGenre
 import com.example.viewmodel.search.searchByKeyword.genre.TvShowGenre
